finished filtering in backend

+++ server/internal/common/response.go
@@ -8,6 +8,6 @@ type ErrorApiResponse struct {
 type OkApiResponse struct {
 	Data        interface{} `json:"data"`
 	StatusCode  int         `json:"statusCode"`
-	CurrentPage int         `json:"currentPage"`
-	TotalPages  int         `json:"totalPages"`
+	CurrentPage int64       `json:"currentPage"`
+	TotalPages  int64       `json:"totalPages"`
 }
